+++ src/routes/tests/issue-227/+page.server.ts
@@ -1,0 +1,15 @@
+import { superValidate } from '$lib/server';
+import { z } from 'zod';
+
+const schema = z.object({
+  name: z.string().default('Hello world!'),
+  email: z.string().email()
+});
+
+export const load = async () => {
+  // Server API:
+  const form = await superValidate(schema);
+
+  // Always return { form } in load and form actions.
+  return { form };
+};